Moved FixedWingVehicle to vehicle folder
